backend/requirements.txt
# ETPS Backend Dependencies

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.0

# Configuration
pyyaml>=6.0

# Data validation
pydantic>=2.0.0
email-validator>=2.0.0

# HTTP and web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
httpx>=0.25.0

# DOCX generation
python-docx>=1.2.0

# TODO: Add additional dependencies as needed
# anthropic>=0.7.0  # Claude API
# openai>=1.0.0     # OpenAI API (fallback)
# qdrant-client>=1.6.0  # Vector store
